Support for add-on groups
This commit adds support for groups that enable arbitrary features for users.
One such example is a Content Module group on EBN. The majority of changes
introduced are to handle IPN requests properly for these groups in order not
to enable/disable user's main subscription but just handle a single feature
group.

Refs #2268.